app_template/init/full/requirements.txt
-r djgram/requirements.txt

# Работа с telegram
aiogram==3.3.0

# Работа с базой данных
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
hiredis==2.3.2

# Прочее
coloredlogs~=15.0.1
python-dotenv==1.0.0
python-json-logger==2.0.7